backend/requirements.txt
selenium==4.9.1
beautifulsoup4==4.12.2
webdriver-manager==4.0.0
lxml
python-dotenv
flask
flask-cors
flask-jwt-extended
werkzeug
pymongo[srv]
clicksend-client